migrate forEach into core API
