+++ global.js
@@ -10,6 +10,7 @@ globalThis.filter = require('./filter')
 globalThis.reduce = require('./reduce')
 globalThis.transform = require('./transform')
 globalThis.flatMap = require('./flatMap')
+globalThis.forEach = require('./forEach')
 globalThis.some = require('./some')
 globalThis.every = require('./every')
 globalThis.and = require('./and')
